requirements.txt
boto3==1.9.207
botocore==1.12.207
celery==5.2.3
django==3.2.16
django-celery-beat==2.2.1
django-celery-results==2.4.0
django-ckeditor==5.7.1
django-configurations==2.3.2
django-extensions==3.1.5
django-jsonfield==1.4.1
django-mysql==3.12.0
django-relative-softdeletion==0.0.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
django-storages==1.11.1
drf-yasg==1.20.0
furl==2.0.0
mysqlclient==2.0.3
neo-mamba==2.0.2
Pillow>=7.1.0
PyJWT==2.4.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.3.0
xlwt==1.3.0
zipp==2.1.0
